Refactors the spec verification tool into its own app
